+++ db.json
@@ -1,0 +1,58 @@
+{
+    "boards": [
+        {
+            "id": 1,
+            "nickName": "피카추",
+            "title": "후기 남겨요 피카츄!",
+            "views": 21,
+            "likes": 24,
+            "diffDate": "1분전",
+            "commentCount": 3
+        },
+        {
+            "id": 2,
+            "nickName": "라이츄",
+            "title": "후기 남겨요 라이츄!",
+            "views": 2,
+            "likes": 31,
+            "diffDate": "30분전",
+            "commentCount": 27
+        },
+        {
+            "id": 3,
+            "nickName": "꼬북이",
+            "title": "후기 남겨요! 꼬북이",
+            "views": 30,
+            "likes": 2,
+            "diffDate": "1분전",
+            "commentCount": 3
+        },
+        {
+            "id": 4,
+            "nickName": "파이리",
+            "title": "후기 남겨요! 파이리",
+            "views": 50,
+            "likes": 100,
+            "diffDate": "1시간전",
+            "commentCount": 3
+        },
+        {
+            "id": 5,
+            "nickName": "이상해씨",
+            "title": "후기 남겨요! 이상해씨",
+            "views": 0,
+            "likes": 0,
+            "diffDate": "1분전",
+            "commentCount": 0
+        },
+        {
+            "id": 6,
+            "nickName": "만나뇽",
+            "title": "후기 남겨요! 만나뇽",
+            "views": 2,
+            "likes": 20,
+            "diffDate": "3분전",
+            "commentCount": 3
+        }
+    ]
+}
