+++ src/components/layout/Header.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled, { css } from 'styled-components';
 import { ContainerDiv } from '../../common/style/common';
 import { HeaderAndFooter } from '../../common/style/common';
-import { Routes, Route, NavLink } from 'react-router-dom';
+import Nav from '../common/Nav';
 
 export default function Header() {
     return (
@@ -12,14 +12,7 @@ export default function Header() {
                     <div className="logo">
                         <img src="/img/logo.svg" alt="LOGO" />
                     </div>
-                    <ul>
-                        <ActiveNavLink to="/">
-                            <li>Home</li>
-                        </ActiveNavLink>
-                        <ActiveNavLink to="/boards">
-                            <li>Community</li>
-                        </ActiveNavLink>
-                    </ul>
+                    <Nav />
                 </div>
                 <ul className="header_right">
                     <li>
@@ -30,38 +23,6 @@ export default function Header() {
         </HeaderWrap>
     );
 }
-const ActiveNavLink = styled(NavLink)`
-    & {
-        position: relative;
-        transition: all 0.3s ease-in-out;
-        &::after {
-            content: '';
-            display: block;
-            position: absolute;
-            left: 0;
-            bottom: -7px;
-            width: 0%;
-            border-bottom: 2px solid #fff;
-            transition: all 0.5s ease-in-out;
-        }
-    }
-
-    &.active {
-        color: #ff4c0e;
-        &:hover {
-            &::after {
-                width: 100%;
-                border-bottom-color: #ff4c0e;
-            }
-        }
-    }
-
-    &:hover {
-        &::after {
-            width: 100%;
-        }
-    }
-`;
 
 const HeaderWrap = styled.div`
     display: flex;
